display selected population boundary on country map
