+++ app/components/maps/BasicMap.js
@@ -160,21 +160,6 @@ class Map extends React.Component {
     );
   }
 
-  addTopoJSONLayer() {
-    L.TopoJSON = L.GeoJSON.extend({
-      addData(jsonData) {
-        if (jsonData.type === 'Topology') {
-          Object.keys(jsonData.objects).forEach((key) => {
-            const geojson = topojson.feature(jsonData, jsonData.objects[key]);
-            L.GeoJSON.prototype.addData.call(this, geojson);
-          });
-        } else {
-          L.GeoJSON.prototype.addData.call(this, jsonData);
-        }
-      }
-    });
-  }
-
   render() {
     return (
       <div id={this.props.id} className="c-map"></div>
